feat: done add/edit category and product

+++ backend/app/Http/Controllers/CategoryController.php
@@ -39,25 +39,46 @@ class CategoryController extends Controller
 
     public function show($id)
     {
-        return Category::find($id);
+        $category = Category::find($id);
+        if ($category === null) {
+            return response(['message' => 'The category does not exist'], 404);
+        }
+
+        return response($category);
     }
 
     public function store(Request $request)
     {
-        return Category::create($request->all());
+        $data = $request->validate([
+            'name' => 'required|string',
+        ]);
+
+        return Category::create($data);
     }
 
     public function update(Request $request, $id)
     {
-        $category = Category::findOrFail($id);
-        $category->update($request->all());
+        $data = $request->validate([
+            'name' => 'required|string',
+        ]);
+
+        $category = Category::find($id);
+        if ($category === null) {
+            return response(['message' => 'The category does not exist'], 404);
+        }
+
+        $category->update($data);
 
         return $category;
     }
 
     public function destroy($id)
     {
-        $category = Category::findOrFail($id);
+        $category = Category::find($id);
+        if ($category === null) {
+            return response(['message' => 'The category does not exist'], 404);
+        }
+
         $category->delete();
 
         return 204;
